requirements.txt
fastapi==0.115.11
aio_pika==9.5.5
aiohttp==3.11.13
aiormq==6.8.1
atomicwrites==1.4.1
CairoSVG==2.7.1
gpiozero==2.0.1
lgpio==0.2.2.0
numpy==1.24.2
picamera2==0.3.25
pigpio==1.78
Pillow==11.1.0
portalocker==3.1.1
psutil==7.0.0
pydantic==2.10.6
pytest==8.3.5
pytz==2022.7.1
PyYAML==6.0.2
pyzipper==0.3.6
qrcode==8.0
raspi_tools==1.0.17
Requests==2.32.3
RPi.GPIO==0.7.1a4
smbus==1.1.post2
smbus2==0.4.2
speedtest_cli==2.1.3
spidev==3.5
tenacity==9.0.0
tinydb==4.8.2
websockets==15.0